refactor(rule): centralize others-if-organic helpers
Move 8.7D-specific helpers into a dedicated module and wire the processor to
use the shared context builders. Add a typed alias for local waste classification
ids to reduce repeated Pick<> usage.
